+++ src/tilemap.h
@@ -42,9 +42,9 @@ public:
 	void width(size_t w);
 	void resize(size_t w, size_t h, Resize_Dialog::Hor_Align ha, Resize_Dialog::Vert_Align va);
 	inline size_t height(void) const { return (_size + _width - 1) / _width; }
-	inline Tile_Tessera *tile(uint8_t x, uint8_t y) const { return _tiles[(size_t)y * _width + (size_t)x]; }
+	inline Tile_Tessera *tile(size_t x, size_t y) const { return _tiles[y * _width + x]; }
 	inline Tile_Tessera *tile(size_t i) const { return _tiles[i]; }
-	inline void tile(uint8_t x, uint8_t y, Tile_Tessera *tt) { _tiles[(size_t)y * _width + (size_t)x] = tt; }
+	inline void tile(size_t x, size_t y, Tile_Tessera *tt) { _tiles[y * _width + x] = tt; }
 	inline void tile(size_t i, Tile_Tessera *tt) { _tiles[i] = tt; }
 	inline Result result(void) const { return _result; }
 	inline bool modified(void) const { return _modified; }
